Efforts to make the project welcoming for others
I enjoy working on the project and love that it's useful to others as well. I
also acknowledge that the project has bugs or not implemented features. To
make the project more accessible, I would like to start an effort to make the
project more welcoming to other contributors of bugs or fixes.

Signed-off-by: Michelangelo van Dam <dragonbe+github@gmail.com>