Fix #76: Removal of auto-prepending of 0 in BE VAT
Issue #76 started the discussion to assist users for bad input by prepending a 0 for Belgian VAT IDs, but since the goal of this library is to validate the correctness of input and the regulation clearly states that Belgian VAT ID's have 10 digits, starting with a "0" we should adhere to the regulation and reject VAT IDs that are still using the older 9-digit VAT IDs without the leading 0.

Since I consider this a bug in the library, I will mark it as such.