Issue 64: Added VIES test service
I have taken a look at this request to access the integration test service of VIES which were made in issue #64 and issue #83. I can conclude that yes, it is possible to access this test service of VIES for integration tests (issue #64), but I was not able to verify if a backend service of VIES was not accessible.

In my test case I used Germany as it was supposed to be "DE-Germany	Unavailable daily from 11:00 PM to 01:30 AM" as was stated on https://ec.europa.eu/taxation_customs/vies/help.html. But on https://ec.europa.eu/taxation_customs/vies/monitoring.html it shows green. Not sure if the information in the help page is still accurate since my normal and test SOAP calls all succeeded.